+++ src/universal/Manager/Interfaces/RunnerManagerInterface.php
@@ -59,23 +59,33 @@ interface RunnerManagerInterface
      * To retrieve a result from an execution, pushed by a runner.
      *
      * @param RunnerInterface $runner
+     * @param TaskInterface $task
      * @param ResultInterface $result
      *
      * @return RunnerManagerInterface
      *
      * @throws \DomainException if the result is not valid for a task registered in the manager
      */
-    public function pushResult(RunnerInterface $runner, ResultInterface $result): RunnerManagerInterface;
+    public function pushResult(
+        RunnerInterface $runner,
+        TaskInterface $task,
+        ResultInterface $result
+    ): RunnerManagerInterface;
 
     /**
      * To allow a runner to update a status of a task.
      *
      * @param RunnerInterface $runner
+     * @param TaskInterface $task
      * @param StatusInterface $status
      *
      * @return RunnerManagerInterface
      */
-    public function pushStatus(RunnerInterface $runner, StatusInterface $status): RunnerManagerInterface;
+    public function pushStatus(
+        RunnerInterface $runner,
+        TaskInterface $task,
+        StatusInterface $status
+    ): RunnerManagerInterface;
 
     /**
      * Called by a runner to inform the manager that it accept to execute the task.
@@ -111,5 +121,8 @@ interface RunnerManagerInterface
      *
      * @throws \DomainException if the task is not executable by the runner
      */
-    public function executeForMeThisTask(TaskManagerInterface $taskManager, TaskInterface $task): RunnerManagerInterface;
+    public function executeForMeThisTask(
+        TaskManagerInterface $taskManager,
+        TaskInterface $task
+    ): RunnerManagerInterface;
 }