+++ src/universal/Runner/RemotePHP7Runner/States/Awaiting.php
@@ -42,7 +42,7 @@ class Awaiting implements StateInterface
 {
     use StateTrait;
 
-    private function doReset()
+    private function doPrepareNextTask()
     {
         /*
          * {@inheritdoc}